Update method signatures and improve code quality in Light.h
This update modifies method signatures in the Light.h file to be more explicit and to follow modern C++ standards. The use of [[nodiscard]] attribute has been added on methods to encourage checking of return values. Optimizations also include changing object passing semantics and removing an unnecessary negative line of code in the namespaces end.