Update material ray intersection condition
The previous implementation caused an issue as it added the material_sample contribution without checking if the material_ray intersected with any primitive in the scene. The updated version corrects this mistake by including an intersect_p check before adding to the L_result. This prevents unjustified contribution from material_sample when there is no actual intersection.