Add BSDFProperties to MaterialSampleResult and update BRDF classes
This commit introduces the inclusion of the "BSDFProperties.h" file and adds a 'properties' object of type BSDFProperties to the MaterialSampleResult structure. The BRDF classes (LambertianBRDF, SpecularReflectionBRDF, and more) have also been updated to return a MaterialSampleResult with proper 'properties' value. Some method signatures and minor behavior changes are also included in this update.