Refactor Tile and TilePixelIterator class constructors
Minor code refactorings are performed in the Tile and TilePixelIterator classes to ensure efficient memory utilization and proper code formatting. A call to `std::move` is added for efficient use of bounds parameter in Tile constructor and indentation in the operator overloading line of TilePixelIterator class has been fixed for better readability.