Refactor Triangle.h to use auto keyword for variable types
The changes made in Triangle.h include the replacement of explicit variable types with the auto keyword to improve readability and make the code cleaner. This has been applied throughout the class including in methods intersect_impl, intersect_p_impl, and others. There is also removal of unnecessary import of the <array> library.